Add foreman setup role
This role aims to do necessary setup in order to make Foreman self
configurable. It's a separate role since need to make sure the installation
has finished and is fully functional. This role is not yet related
to foreman_setup plugin or any of its successor but once that's available
it should probably configure it too.